backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
redis
weaviate-client
google-generativeai
python-dotenv
httpx
langchain
langgraph
langchain-google-genai
playwright
